first refactoring. more to go
