+++ VISAP/html/trawling.html
@@ -6,8 +6,8 @@
   <h2 class="h3">Captura en biomassa per port</h2>
   <div class="btn-toolbar mb-2 mb-mb-0"><!--class="btn-group me-2"-->
     <div class="btn-group me-2">
-      <button type="button" id="filterIsOnBtn" onclick="deactivateFilter(event)" class="btn btn-sm btn-danger" style="visibility:hidden">✖ Desactivar filter</button>
-      <button type="button" id="filterSpeciesBtn" onclick="filterSpeciesGUI(event)" w3-include-html="filter.html" class="btn btn-sm btn-outline-secondary"><span data-feather="filter"></span> Filtrar per espècie</button>
+      <button type="button" id="filterIsOnBtn" onclick="deactivateFilter(event, 'portBiomass')" class="btn btn-sm btn-danger" style="visibility:hidden">✖ Desactivar filter</button>
+      <button type="button" id="filterSpeciesBtn" onclick="filterSpeciesGUI(event, 'portBiomass')" w3-include-html="filter.html" class="btn btn-sm btn-outline-secondary"><span data-feather="filter"></span> Filtrar per espècie</button>
     </div>
     <div class="dropdown">
       <button type="button" class="btn btn-sm btn-outline-secondary dropdown-toggle" id="dropdownMenuButton" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
@@ -15,8 +15,8 @@
         Exportar
       </button>
       <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
-        <a class="dropdown-item" onclick="exportCSV(event)">.csv</a>
-        <a class="dropdown-item" onclick="exportJSON(event)">.json</a>
+        <a class="dropdown-item" onclick="exportCSV(event, 'pesca_arrossegament_port_biomassa')">.csv</a>
+        <a class="dropdown-item" onclick="exportJSON(event, 'pesca_arrossegament_port_biomassa')">.json</a>
       </div>
     </div>
   </div>
@@ -25,12 +25,12 @@
 <!-- Pie chart -->
 <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-3 pb-2 mb-3 border-bottom">
   <!-- Class will change when loaded -->
-  <div class="loader mx-auto" id="piechart" width="600" height="600" style="max-width: 600px;"></div>
+  <div class="loader mx-auto" id="piePortsBiomass" width="600" height="600" style="max-width: 600px;"></div>
   <!--div class="mx-auto" id="comparePie" width="500" height="500" style="max-width: 500px;"></div-->
-  <button type="button" id="compareBtn" onclick="compareTrawling(event)" class="btn btn-sm btn-outline-secondary mx-auto"><span data-feather="plus-circle"></span> Comparar</button>
-  <button type="button" id="closeCompare" onclick="closeCompare(event)" style="visibility:hidden;" class="btn btn-sm btn-outline-secondary mx-auto">Tancar comparació</button>
+  <button type="button" id="compareBtn" onclick="compareTrawling(event, 'portBiomass')" class="btn btn-sm btn-outline-secondary mx-auto"><span data-feather="plus-circle"></span> Comparar</button>
+  <button type="button" id="closeCompare" onclick="closeCompare(event, 'portBiomass')" style="visibility:hidden;" class="btn btn-sm btn-outline-secondary mx-auto">Tancar comparació</button>
   <!-- Overlay functions are defined in createHTML.js-->
-  <div id="overlay" style="visibility:hidden" onclick="closeFilterGUI(event)">
+  <div id="overlay" style="visibility:hidden" onclick="closeFilterGUI(event, 'portBiomass')">
     <div id="info"></div>
   </div>
 </div>
@@ -49,8 +49,8 @@
         Exportar
       </button>
       <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
-        <a class="dropdown-item" onclick="exportCSV(event)">.csv</a>
-        <a class="dropdown-item" onclick="exportJSON(event)">.json</a>
+        <a class="dropdown-item" onclick="exportCSV(event, 'pesca_arrossegament_any_biomassa')">.csv</a>
+        <a class="dropdown-item" onclick="exportJSON(event, 'pesca_arrossegament_any_biomassa')">.json</a>
       </div>
     </div>
   </div>
